+++ joe/options.c
@@ -28,56 +28,56 @@ OPTIONS pdefault = {
 	NULL,		/* *next */
 	"prompt",	/* ftype */
 	NULL,		/* *match */
-	0,		/* overtype */
 	0,		/* lmargin */
 	76,		/* rmargin */
-	0,		/* autoindent */
-	0,		/* wordwrap */
-	0,		/* nobackup */
-	8,		/* tab */
+	false,		/* overtype */
+	false,		/* autoindent */
+	false,		/* wordwrap */
+	false,		/* nobackup */
 	' ',		/* indent char */
+	8,		/* tab */
 	1,		/* indent step */
 	NULL,		/* *context */
 	NULL,		/* *lmsg */
 	NULL,		/* *rmsg */
 	NULL,		/* *smsg */
 	NULL,		/* *zmsg */
-	0,		/* line numbers */
-	0,		/* highlight current line */
-	0,		/* read only */
-	0,		/* french spacing */
-	0,		/* flowed text */
-	0,		/* spaces */
-#ifdef __MSDOS__
-	1,		/* crlf */
-#else
-	0,		/* crlf */
-#endif
-	0,		/* Highlight */
-	0,		/* Visible whitespace */
-	0,		/* Syntax debugging */
 	NULL,		/* Syntax name */
 	NULL,		/* Syntax */
 	NULL,		/* Name of character set */
 	NULL,		/* Character set */
 	NULL,		/* Language */
-	0,		/* Smart home key */
-	0,		/* Goto indent first */
-	0,		/* Smart backspace key */
-	0,		/* Purify indentation */
-	0,		/* Picture mode */
-	0,		/* highlighter_context */
-	0,		/* single_quoted */
-	0,		/* no_double_quoted */
-	0,		/* c_comment */
-	0,		/* cpp_comment */
-	0,		/* hash_comment */
-	0,		/* vhdl_comment */
-	0,		/* semi_comment */
-	0,		/* tex_comment */
-	0,		/* hex */
-	0,		/* hide ansi */
-	0,		/* status line context */
+	false,		/* line numbers */
+	false,		/* highlight current line */
+	false,		/* read only */
+	false,		/* french spacing */
+	false,		/* flowed text */
+	false,		/* spaces */
+#ifdef __MSDOS__
+	true,		/* crlf */
+#else
+	false,		/* crlf */
+#endif
+	false,		/* Highlight */
+	false,		/* Visible whitespace */
+	false,		/* Syntax debugging */
+	false,		/* Smart home key */
+	false,		/* Goto indent first */
+	false,		/* Smart backspace key */
+	false,		/* Purify indentation */
+	false,		/* Picture mode */
+	false,		/* highlighter_context */
+	false,		/* single_quoted */
+	false,		/* no_double_quoted */
+	false,		/* c_comment */
+	false,		/* cpp_comment */
+	false,		/* hash_comment */
+	false,		/* vhdl_comment */
+	false,		/* semi_comment */
+	false,		/* tex_comment */
+	false,		/* hex */
+	false,		/* hide ansi */
+	false,		/* status line context */
 	NULL,		/* text_delimiters */
 	NULL,		/* Characters which can indent paragraphs */
 	NULL,		/* Characters which begin non-paragraph lines */
@@ -94,56 +94,56 @@ OPTIONS fdefault = {
 	NULL,		/* *next */
 	"default",	/* ftype */
 	NULL,		/* *match */
-	0,		/* overtype */
 	0,		/* lmargin */
 	76,		/* rmargin */
-	0,		/* autoindent */
-	0,		/* wordwrap */
-	0,		/* nobackup */
-	8,		/* tab */
+	false,		/* overtype */
+	false,		/* autoindent */
+	false,		/* wordwrap */
+	false,		/* nobackup */
 	' ',		/* indent char */
+	8,		/* tab */
 	1,		/* indent step */
 	"main",		/* *context */
 	"\\i%n %m %M",	/* *lmsg */
 	" %S Ctrl-K H for help",	/* *rmsg */
 	NULL,		/* *smsg */
 	NULL,		/* *zmsg */
-	0,		/* line numbers */
-	0,		/* higlight current line */
-	0,		/* read only */
-	0,		/* french spacing */
-	0,		/* flowed text */
-	0,		/* spaces */
-#ifdef __MSDOS__
-	1,		/* crlf */
-#else
-	0,		/* crlf */
-#endif
-	0,		/* Highlight */
-	0,		/* Visible whitespace */
-	0,		/* Syntax debugging */
 	NULL,		/* Syntax name */
 	NULL,		/* Syntax */
 	NULL,		/* Name of character set */
 	NULL,		/* Character set */
 	NULL,		/* Language */
-	0,		/* Smart home key */
-	0,		/* Goto indent first */
-	0,		/* Smart backspace key */
-	0,		/* Purity indentation */
-	0,		/* Picture mode */
-	0,		/* highlighter_context */
-	0,		/* single_quoted */
-	0,		/* no_double_quoted */
-	0,		/* c_comment */
-	0,		/* cpp_comment */
-	0,		/* hash_comment */
-	0,		/* vhdl_comment */
-	0,		/* semi_comment */
-	0,		/* tex_comment */
-	0,		/* hex */
-	0,		/* hide ansi */
-	0,		/* status line context */
+	false,		/* line numbers */
+	false,		/* higlight current line */
+	false,		/* read only */
+	false,		/* french spacing */
+	false,		/* flowed text */
+	false,		/* spaces */
+#ifdef __MSDOS__
+	true,		/* crlf */
+#else
+	false,		/* crlf */
+#endif
+	false,		/* Highlight */
+	false,		/* Visible whitespace */
+	false,		/* Syntax debugging */
+	false,		/* Smart home key */
+	false,		/* Goto indent first */
+	false,		/* Smart backspace key */
+	false,		/* Purity indentation */
+	false,		/* Picture mode */
+	false,		/* highlighter_context */
+	false,		/* single_quoted */
+	false,		/* no_double_quoted */
+	false,		/* c_comment */
+	false,		/* cpp_comment */
+	false,		/* hash_comment */
+	false,		/* vhdl_comment */
+	false,		/* semi_comment */
+	false,		/* tex_comment */
+	false,		/* hex */
+	false,		/* hide ansi */
+	false,		/* status line context */
 	NULL,		/* text_delimiters */
 	">;!#%/",	/* Characters which can indent paragraphs */
 	".",	/* Characters which begin non-paragraph lines */
@@ -201,41 +201,43 @@ void lazy_opts(B *b, OPTIONS *o)
 	if (!b->o.language)
 		b->o.language = locale_msgs;
 	if (b->o.hex) {
+		b->o.hex_saved = 0;
+
 		/* Hex not allowed with UTF-8 */
 		if (b->o.charmap->type) {
 			b->o.charmap = find_charmap("c");
-			b->o.hex |= HEX_RESTORE_UTF8;
+			b->o.hex_saved |= HEX_RESTORE_UTF8;
 		}
 
 		/* Hex not allowed with CRLF */
 		if (b->o.crlf) {
 			b->o.crlf = 0;
-			b->o.hex |= HEX_RESTORE_CRLF;
+			b->o.hex_saved |= HEX_RESTORE_CRLF;
 		}
 
 		if (!b->o.overtype) {
 			b->o.overtype = 1;
-			b->o.hex |= HEX_RESTORE_INSERT;
+			b->o.hex_saved |= HEX_RESTORE_INSERT;
 		}
 
 		if (b->o.wordwrap) {
 			b->o.wordwrap = 0;
-			b->o.hex |= HEX_RESTORE_WORDWRAP;
+			b->o.hex_saved |= HEX_RESTORE_WORDWRAP;
 		}
 
 		if (b->o.autoindent) {
 			b->o.autoindent = 0;
-			b->o.hex |= HEX_RESTORE_AUTOINDENT;
+			b->o.hex_saved |= HEX_RESTORE_AUTOINDENT;
 		}
 
 		if (b->o.ansi) {
 			b->o.ansi = 0;
-			b->o.hex |= HEX_RESTORE_ANSI;
+			b->o.hex_saved |= HEX_RESTORE_ANSI;
 		}
 
 		if (b->o.picture) {
 			b->o.picture = 0;
-			b->o.hex |= HEX_RESTORE_PICTURE;
+			b->o.hex_saved |= HEX_RESTORE_PICTURE;
 		}
 	}
 
@@ -300,7 +302,7 @@ enum opt_type {
 union opt_storage_p {
 	void  *v;
 	char  *c;
-	int   *b; /* bool */
+	bool  *b;
 	int   *i;
 	off_t *o;
 	char **s;
@@ -704,7 +706,7 @@ int glopt(char *s, char *arg, OPTIONS *options, int set)
 			break;
 		case LOC_OPT_BOOL: /* Local option flag */
 			if (options)
-				*OPTPTR(options, opt->ofst, int) = st;
+				*OPTPTR(options, opt->ofst, bool) = st;
 			ret = 1;
 			break;
 		case LOC_OPT_INT: /* Local option integer */
@@ -1258,9 +1260,9 @@ static int find_option(char *s)
 	return -1;
 }
 
-static int applyopt(BW *bw, int *optp, int y, int flg)
+static int applyopt(BW *bw, bool *optp, int y, int flg)
 {
-	int oldval, newval;
+	bool oldval, newval;
 	const char *msg;
 
 	oldval = *optp;
@@ -1302,7 +1304,7 @@ static int olddoopt(BW *bw, int y, int flg, int *notify)
 			applyopt(bw, glopts[y].set.b, y, flg);
 			break;
 		case LOC_OPT_BOOL:
-			oldval = applyopt(bw, OPTPTR(&bw->o, glopts[y].ofst, int), y, flg);
+			oldval = applyopt(bw, OPTPTR(&bw->o, glopts[y].ofst, bool), y, flg);
 
 			/* Propagate readonly bit to B */
 			if (glopts[y].ofst == offsetof(OPTIONS, readonly))
@@ -1310,43 +1312,46 @@ static int olddoopt(BW *bw, int y, int flg, int *notify)
 
 			/* Kill UTF-8 and CRLF modes if we switch to hex display */
 			if (glopts[y].ofst == offsetof(OPTIONS, hex)) {
+				oldval = bw->o.hex_saved;
+
 				if (bw->o.hex && !oldval) {
 					bw->o.hex = 1;
+					bw->o.hex_saved = 0;
 					if (bw->b->o.charmap->type) {
 						/* Switch out of UTF-8 mode */
 						doencoding(bw->parent, vsncpy(NULL, 0, sc("C")), NULL, NULL);
-						bw->o.hex |= HEX_RESTORE_UTF8;
+						bw->o.hex_saved |= HEX_RESTORE_UTF8;
 					}
 
 					if (bw->o.crlf) {
 						/* Switch out of CRLF mode */
 						bw->o.crlf = 0;
-						bw->o.hex |= HEX_RESTORE_CRLF;
+						bw->o.hex_saved |= HEX_RESTORE_CRLF;
 					}
 
 					if (!bw->o.overtype) {
 						bw->o.overtype = 1;
-						bw->o.hex |= HEX_RESTORE_INSERT;
+						bw->o.hex_saved |= HEX_RESTORE_INSERT;
 					}
 
 					if (bw->o.wordwrap) {
 						bw->o.wordwrap = 0;
-						bw->o.hex |= HEX_RESTORE_WORDWRAP;
+						bw->o.hex_saved |= HEX_RESTORE_WORDWRAP;
 					}
 
 					if (bw->o.autoindent) {
 						bw->o.autoindent = 0;
-						bw->o.hex |= HEX_RESTORE_AUTOINDENT;
+						bw->o.hex_saved |= HEX_RESTORE_AUTOINDENT;
 					}
 
 					if (bw->o.ansi) {
 						bw->o.ansi = 0;
-						bw->o.hex |= HEX_RESTORE_ANSI;
+						bw->o.hex_saved |= HEX_RESTORE_ANSI;
 					}
 
 					if (bw->o.picture) {
 						bw->o.picture = 0;
-						bw->o.hex |= HEX_RESTORE_PICTURE;
+						bw->o.hex_saved |= HEX_RESTORE_PICTURE;
 					}
 					/* Try to put entire hex dump on screen in case where we were
 					   scrolled far to the right */
@@ -1504,7 +1509,7 @@ const char *get_status(BW *bw, char *s)
 				joe_snprintf_1(buf, OPT_BUF_SIZE, "%s", *glopts[y].set.s ? *glopts[y].set.s : "");
 				return buf;
 			case LOC_OPT_BOOL:
-				return *OPTPTR(&bw->o, glopts[y].ofst, int) ? "ON" : "OFF";
+				return *OPTPTR(&bw->o, glopts[y].ofst, bool) ? "ON" : "OFF";
 			case LOC_OPT_INT:
 				joe_snprintf_1(buf, OPT_BUF_SIZE, "%d", *OPTPTR(&bw->o, glopts[y].ofst, int));
 				return buf;