+++ joe/ublock.c
@@ -15,8 +15,8 @@ int nowmarking;
 
 /* Global options */
 
-int square = 0;			/* Set for rectangle mode */
-int lightoff = 0;		/* Set if highlighting should turn off
+bool square = 0;			/* Set for rectangle mode */
+bool lightoff = 0;		/* Set if highlighting should turn off
 
 				   after block operations */
 /* Global variables */
@@ -75,7 +75,7 @@ int upop(W *w, int k)
 /* Return true if markb/markk are valid */
 /* If r is set, swap markb with markk if necessary */
 
-int autoswap;
+bool autoswap;
 
 int markv(int r)
 {
@@ -1221,7 +1221,7 @@ static int dofilt(W *w, char *s, void *object, int *notify)
 					       markk->line - markb->line + 1,
 					       markk->xcol);
 
-			tmp = bread(fr[0], MAXOFF);
+			tmp = bread(fr[0], MAXOFF, 0);
 			if (piscol(tmp->eof))
 				height = tmp->eof->line + 1;
 			else
@@ -1248,7 +1248,7 @@ static int dofilt(W *w, char *s, void *object, int *notify)
 			if (!flg)
 				prgetc(p);
 			bdel(markb, p);
-			binsb(p, bread(fr[0], MAXOFF));
+			binsb(p, bread(fr[0], MAXOFF, 0));
 			if (!flg) {
 				pset(p,markk);
 				prgetc(p);